Day 5: Add task status and timestamp, implement filtering, sorting, and status update

+++ app/routes/tasks.py
@@ -1,10 +1,11 @@
-from fastapi import APIRouter,HTTPException,Depends
+from fastapi import APIRouter,HTTPException,Depends,status
 from typing import List
-from app.models import *
+from app.models import Task
 from uuid import UUID,uuid4
-from app.schema import TaskCreate,TaskResponse,Tasklistresponse,Taskcreateresponse
+from app.schema import *
 from app.database import SessionLocal
 from sqlalchemy.orm import Session
+from app.exception import TaskNotFoundExpection
 
 router=APIRouter()
 tasks_db : List[Task_response]=[]
@@ -17,7 +18,7 @@ def get_db():
     finally:
         db.close()
 
-@router.post("/create",response_model=Taskcreateresponse)
+@router.post("/create",response_model=Taskcreateresponse, status_code=status.HTTP_201_CREATED)
 def create_task(task : TaskCreate, db :Session =Depends(get_db)):
     task_id=str(uuid4())
     db_task= Task(id=task_id, name=task.name, description=task.description)
@@ -26,38 +27,43 @@ def create_task(task : TaskCreate, db :Session =Depends(get_db)):
     db.refresh(db_task)
     return {"message" : "hey chetan , created successfully"}
 
-@router.get("/tasks", response_model=Tasklistresponse)
+@router.get("/tasks", response_model=Tasklistresponse ,status_code=status.HTTP_200_OK)
 def get_all_tasks(db :Session =Depends(get_db)):
     tasks_db=db.query(Task).all()
     return{ "message":"all tasks fetched",
            "data": tasks_db}
 
-
-@router.get("/tasks/{task_id}", response_model=TaskbyID)
+@router.get("/tasks/{task_id}", response_model=TaskbyID ,status_code=status.HTTP_200_OK)
 def get_task_by_id(task_id :str , db :Session =Depends(get_db)):
     tasks_db=db.query(Task).all()
-    for task in tasks_db:
-        if task.id==task_id:
-            return { "message":"data found","data":task}
-        
-    raise HTTPException(status_code=404, detail="Task not found")
 
-@router.delete("/delete/{task_id}", response_model=Tasklistresponse)
+    if not tasks_db:
+        raise TaskNotFoundExpection
+
+    else:
+        for task in tasks_db:
+            if task.id==task_id:
+                return { "message":"data found","data":task}
+
+@router.delete("/delete/{task_id}", status_code=status.HTTP_204_NO_CONTENT)
 def Delete_task(task_id :str , db :Session =Depends(get_db)):
     tasks_db=db.query(Task).all()
-    for task in tasks_db:
-        if task.id==task_id:
-            db.delete(task)
-            db.commit()
-            db.refresh(tasks_db)
-            return { "message":"Task deleted successfully","data":tasks_db}
-        
-    raise HTTPException(status_code=404, detail="Task not found")
 
+    if not tasks_db:
+        raise TaskNotFoundExpection
+    
+    else :
+        for task in tasks_db:
+            if task.id==task_id:
+                db.delete(task)
+                db.commit()
+                db.refresh(tasks_db)
+                return { "message":"Task deleted successfully"}
+        
 @router.put("/update/{id}", response_model=TaskResponse)
 def update_task(id: str, updated_task: TaskCreate, db: Session = Depends(get_db)):
     task = db.query(Task).filter(Task.id == id).first()
-    print(task)
+    
     if not task:
         raise HTTPException(status_code=404, detail="Task not found")
 
@@ -67,3 +73,30 @@ def update_task(id: str, updated_task: TaskCreate, db: Session = Depends(get_db)
     db.refresh(task)
 
     return task
+
+@router.post("/create/bulk", response_model= Tasklistresponse ,status_code=status.HTTP_201_CREATED)
+def Bulk_create(tasks : List[TaskCreate],db :Session =Depends(get_db)):
+    for task in tasks:
+        task_id=str(uuid4())
+        db_task= Task(id=task_id, name=task.name, description=task.description)
+        db.add(db_task)
+        db.commit()
+        db.refresh(db_task)
+        
+    task_db = db.query(Task).all()
+    return{ "message":"all tasks fetched",
+           "data":task_db }
+
+@router.put("update/status/{id}", response_model=Taskupdateresponse,status_code=status.HTTP_200_OK)
+def update_status(id : str , db :Session =Depends(get_db) ):
+    task =db.query(Task).filter(Task.id==id).first()
+    print(task)
+    if not task:
+        raise TaskNotFoundExpection
+    
+    else :
+        task.completed = 1
+        db.commit()
+        db.refresh(task)
+    
+    return {"message" : "status updated successfully", "data":task}
